// src/controller/contact-controller.ts
import type { Response, NextFunction } from 'express';
import type { UserRequest } from '../type/user-request';
import type { CreateContactRequest, SearchContactRequest, UpdateContactRequest } from '../model/contact-model';
import { ContactService } from '../service/contact-service';
import { logger } from '../application/logging';

export class ContactController {
  static async create(req: UserRequest, res: Response, next: NextFunction) {
    try {
      const request: CreateContactRequest = req.body as CreateContactRequest;
      const result = await ContactService.create(req.user!, request);
      res.status(201).json({
        status: 'success',
        data: result,
      });
    } catch (error) {
      next(error);
    }
  }

  static async get(req: UserRequest, res: Response, next: NextFunction) {
    try {
      const contactId = Number(req.params.contactId);
      const result = await ContactService.get(req.user!, contactId);
      res.status(200).json({
        status: 'success',
        data: result,
      });
    } catch (error) {
      next(error);
    }
  }

  static async update(req: UserRequest, res: Response, next: NextFunction) {
    try {
      const request: UpdateContactRequest = req.body as UpdateContactRequest;
      request.id = Number(req.params.contactId);

      const result = await ContactService.update(req.user!, request);
      res.status(200).json({
        status: 'success',
        data: result,
      });
    } catch (error) {
      next(error);
    }
  }

  static async remove(req: UserRequest, res: Response, next: NextFunction) {
    try {
      const contactId = Number(req.params.contactId);
      await ContactService.remove(req.user!, contactId);
      res.status(200).json({
        status: 'success',
        data: {
          message: 'delete success',
        },
      });
    } catch (error) {
      next(error);
    }
  }

  static async search(req: UserRequest, res: Response, next: NextFunction) {
    try {
      const request: SearchContactRequest = {
        name: req.query.name as string,
        email: req.query.email as string,
        phone: req.query.phone as string,
        page: req.query.page ? Number(req.query.page) : 1,
        size: req.query.size ? Number(req.query.size) : 10,
      };

      const { data, paging } = await ContactService.search(req.user!, request);
      res.status(200).json({
        status: 'success',
        data,
        paging,
      });
    } catch (error) {
      next(error);
    }
  }
}
